+++ lib/src/model/package_graph.dart
@@ -753,7 +753,8 @@ class PackageGraph with CommentReferable, Nameable, ModelBuilder {
   /// This doesn't know anything about [PackageGraph.inheritThrough] and probably
   /// shouldn't, so using it with [Inheritable]s without special casing is
   /// not advised.
-  ModelElement findCanonicalModelElementFor(Element e,
+  // FIXME(nnbd): remove null check ignore in model_utils after migration
+  ModelElement /*?*/ findCanonicalModelElementFor(Element e,
       {Container preferredClass}) {
     assert(allLibrariesAdded);
     var lib = findCanonicalLibraryFor(e);
